HRPV2-29 fixes for PR, improve legibility and remove commented code

+++ oe-hr-portal-service-model/src/main/java/com/openenglish/hr/persistence/repository/PersonCourseSummaryRepository.java
@@ -8,7 +8,5 @@ import java.util.List;
 
 public interface PersonCourseSummaryRepository extends JpaRepository<PersonCourseSummary, Long> {
     List<PersonCourseSummary> findPersonCourseSummaryByPersonDetailsSalesforcePurchaserId(String salesforcePurchaserId);
-    List<PersonCourseSummary> findPersonCourseSummaryByPersonDetailsSalesforcePurchaserIdAndCreatedDateBetweenAndCourseCourseTypeIdIn(String salesforcePurchaserId, LocalDateTime startDate, LocalDateTime endDate, List<Integer> courseTypes);
-//    List<PersonCourseSummary> findPersonCourseSummaryByCourseCourseTypeIdIn(List<Long> courseTypes);
-//List<PersonCourseSummary> findPersonCourseSummaryByPersonDetailsSalesforcePurchaserIdAndCourseCourseTypeIdIn(String salesforcePurchaserId,List<Long> courseTypes);
+    List<PersonCourseSummary> findPersonCourseSummaryByPersonDetailsSalesforcePurchaserIdAndCreatedDateBetweenAndCourseCourseTypeIdIn(String salesforcePurchaserId, LocalDateTime startDate, LocalDateTime endDate, List<Long> courseTypes);
 }
