+++ oe-hr-portal-service-service/src/test/java/com/openenglish/hr/service/ActivityServiceTest.java
@@ -32,37 +32,37 @@ public class ActivityServiceTest {
         String salesforcePurchaserId = "12345";
 
         PersonCourseSummary personCourseSummary11 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(1).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(1L).build()).build())
                 .timeontask(50)
                 .build();
         PersonCourseSummary personCourseSummary12 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(1).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(1L).build()).build())
                 .timeontask(70)
                 .build();
         PersonCourseSummary personCourseSummary13 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(1).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(1L).build()).build())
                 .timeontask(90)
                 .build();
 
         PersonCourseSummary personCourseSummary21 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(3).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(3L).build()).build())
                 .timeontask(10)
                 .build();
         PersonCourseSummary personCourseSummary22 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(3).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(3L).build()).build())
                 .timeontask(10)
                 .build();
 
         PersonCourseSummary personCourseSummary31 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(4).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(4L).build()).build())
                 .timeontask(30)
                 .build();
         PersonCourseSummary personCourseSummary32 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(4).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(4L).build()).build())
                 .timeontask(30)
                 .build();
         PersonCourseSummary personCourseSummary33 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(4).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(4L).build()).build())
                 .timeontask(30)
                 .build();
 
@@ -129,44 +129,44 @@ public class ActivityServiceTest {
         final int MONTHS_OF_YEAR = 12;
 
         PersonCourseSummary personCourseSummary11 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(1).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(1L).build()).build())
                 .createdDate(LocalDateTime.of(2022, 1, 15, 12, 0, 0))
                 .timeontask(50)
                 .build();
         PersonCourseSummary personCourseSummary12 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(1).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(1L).build()).build())
                 .createdDate(LocalDateTime.of(2022, 1, 19, 14, 0, 0))
                 .timeontask(70)
                 .build();
         PersonCourseSummary personCourseSummary13 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(1).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(1L).build()).build())
                 .createdDate(LocalDateTime.of(2022, 1, 20, 12, 0, 0))
                 .timeontask(90)
                 .build();
 
         PersonCourseSummary personCourseSummary21 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(3).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(3L).build()).build())
                 .createdDate(LocalDateTime.of(2022, 1, 21, 10, 0, 0))
                 .timeontask(10)
                 .build();
         PersonCourseSummary personCourseSummary22 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(3).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(3L).build()).build())
                 .createdDate(LocalDateTime.of(2022, 2, 15, 12, 0, 0))
                 .timeontask(10)
                 .build();
 
         PersonCourseSummary personCourseSummary31 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(4).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(4L).build()).build())
                 .createdDate(LocalDateTime.of(2022, 2, 7, 8, 0, 0))
                 .timeontask(30)
                 .build();
         PersonCourseSummary personCourseSummary32 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(4).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(4L).build()).build())
                 .createdDate(LocalDateTime.of(2022, 2, 7, 12, 0, 0))
                 .timeontask(30)
                 .build();
         PersonCourseSummary personCourseSummary33 = PersonCourseSummary.builder()
-                .course(Course.builder().courseType(CourseType.builder().id(4).build()).build())
+                .course(Course.builder().courseType(CourseType.builder().id(4L).build()).build())
                 .createdDate(LocalDateTime.of(2022, 3, 1, 12, 0, 0))
                 .timeontask(30)
                 .build();
@@ -175,11 +175,11 @@ public class ActivityServiceTest {
                 personCourseSummary21, personCourseSummary22, personCourseSummary31, personCourseSummary32, personCourseSummary33);
 
         new Expectations() {{
-            personCourseSummaryRepository.findPersonCourseSummaryByPersonDetailsSalesforcePurchaserIdAndCreatedDateBetweenAndCourseCourseTypeIdIn(anyString, (LocalDateTime) any, (LocalDateTime) any, (List<Integer>) any);
+            personCourseSummaryRepository.findPersonCourseSummaryByPersonDetailsSalesforcePurchaserIdAndCreatedDateBetweenAndCourseCourseTypeIdIn(anyString, (LocalDateTime) any, (LocalDateTime) any, (List<Long>) any);
             returns(personCourseSummaries);
         }};
 
-        List<ActivityStatistics> activityStatistics = activityService.getActivitiesStatistics(salesforcePurchaserId, 0, List.of(1, 2));
+        List<ActivityStatistics> activityStatistics = activityService.getActivitiesStatistics(salesforcePurchaserId, 0, List.of(1L, 2L));
         assertEquals(MONTHS_OF_YEAR, activityStatistics.size());
         activityStatistics
                 .stream()
@@ -197,11 +197,11 @@ public class ActivityServiceTest {
         List<PersonCourseSummary> personCourseSummaries = new ArrayList<>();
 
         new Expectations() {{
-            personCourseSummaryRepository.findPersonCourseSummaryByPersonDetailsSalesforcePurchaserIdAndCreatedDateBetweenAndCourseCourseTypeIdIn(anyString, (LocalDateTime) any, (LocalDateTime) any, (List<Integer>) any);
+            personCourseSummaryRepository.findPersonCourseSummaryByPersonDetailsSalesforcePurchaserIdAndCreatedDateBetweenAndCourseCourseTypeIdIn(anyString, (LocalDateTime) any, (LocalDateTime) any, (List<Long>) any);
             returns(personCourseSummaries);
         }};
 
-        List<ActivityStatistics> activityStatistics = activityService.getActivitiesStatistics(salesforcePurchaserId, 0, List.of(1, 2));
+        List<ActivityStatistics> activityStatistics = activityService.getActivitiesStatistics(salesforcePurchaserId, 0, List.of(1L, 2L));
         assertEquals(MONTHS_OF_YEAR, activityStatistics.size());
         activityStatistics
                 .stream()
